chore: refactor & cleanup
