+++ src/components/chain-token-selector/token-selector.tsx
@@ -16,16 +16,33 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog";
+import {
+  Drawer,
+  DrawerTrigger,
+  DrawerHeader,
+  DrawerTitle,
+  DrawerContent,
+} from "@/components/ui/drawer";
 import { Skeleton } from "@/components/ui/skeleton";
 import { useEvmTokenBalances } from "@/lib/hooks/useTokenBalances";
-import { cn, formatNumber, formatTokenAmount, tokenKey } from "@/lib/utils";
+import {
+  cn,
+  DESKTOP_MEDIA_QUERY,
+  formatNumber,
+  formatTokenAmount,
+  tokenKey,
+} from "@/lib/utils";
 import { ChevronsUpDown } from "lucide-react";
 import { useState } from "react";
 import { ChainTokenLogo } from "../chain-token-logo";
 import { TokenListItemProps, TokenSelectorProps } from "./types";
+import { useMediaQuery } from "@/lib/hooks/useMediaQuery";
+import { Token } from "@/types";
 
 const TokenSelector: React.FC<TokenSelectorProps> = (props) => {
-  const [open, setOpen] = useState(false);
+  const isDesktop = useMediaQuery(DESKTOP_MEDIA_QUERY);
+
+  const [open, setOpen] = useState<boolean>(false);
 
   const { data, isLoading: isBalancesLoading } = useEvmTokenBalances(
     props.wallet as `0x${string}`,
@@ -36,9 +53,44 @@ const TokenSelector: React.FC<TokenSelectorProps> = (props) => {
   );
   const balances = data?.balances;
 
+  if (isDesktop) {
+    return (
+      <Dialog open={open} onOpenChange={setOpen}>
+        <DialogTrigger asChild>
+          <Button
+            variant="outline"
+            role="combobox"
+            aria-expanded={open}
+            className="w-[200px] justify-between text-xs"
+          >
+            <div className="inline-flex items-center gap-2">
+              {props.token?.address && <ChainTokenLogo token={props.token} />}
+              {props.token ? props.token.symbol : "Select token..."}
+            </div>
+            <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
+          </Button>
+        </DialogTrigger>
+
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle className="text-sm">Select a token</DialogTitle>
+          </DialogHeader>
+
+          <TokenSelection
+            isBalancesLoading={isBalancesLoading}
+            balances={balances}
+            selectedToken={props.token}
+            onSelectedTokenChange={props.onTokenChange}
+            setOpen={setOpen}
+          />
+        </DialogContent>
+      </Dialog>
+    );
+  }
+
   return (
-    <Dialog open={open} onOpenChange={setOpen}>
-      <DialogTrigger asChild>
+    <Drawer open={open} onOpenChange={setOpen}>
+      <DrawerTrigger asChild>
         <Button
           variant="outline"
           role="combobox"
@@ -51,58 +103,86 @@ const TokenSelector: React.FC<TokenSelectorProps> = (props) => {
           </div>
           <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
         </Button>
-      </DialogTrigger>
+      </DrawerTrigger>
 
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle className="text-sm">Select a token</DialogTitle>
-        </DialogHeader>
+      <DrawerContent>
+        <DrawerHeader>
+          <DrawerTitle className="text-sm">Select a token</DrawerTitle>
+        </DrawerHeader>
 
-        <Command
-          defaultValue="-"
-          className="min-h-[340px] max-h-[45svh] bg-transparent border"
-        >
-          <CommandInput placeholder="Search token..." className="text-xs" />
-          <CommandList>
-            {!isBalancesLoading && (
-              <CommandEmpty className="text-xs text-center py-3">
-                No token found.
-              </CommandEmpty>
-            )}
+        <div className="px-4 pb-4">
+          <TokenSelection
+            isBalancesLoading={isBalancesLoading}
+            balances={balances}
+            selectedToken={props.token}
+            onSelectedTokenChange={props.onTokenChange}
+            setOpen={setOpen}
+          />
+        </div>
+      </DrawerContent>
+    </Drawer>
+  );
+};
 
-            {isBalancesLoading ? (
-              <div className="p-2">
-                {Array.from({ length: 7 }).map((_, i) => (
-                  <div key={i} className="flex items-center gap-2 p-2">
-                    <Skeleton className="h-6 min-w-6 rounded-full" />
-                    <div className="w-full flex items-center justify-between">
-                      <Skeleton className="h-4 w-14" />
-                      <Skeleton className="h-4 w-24" />
-                    </div>
-                  </div>
-                ))}
+const TokenSelection: React.FC<{
+  isBalancesLoading: boolean;
+  balances: Array<Token> | undefined;
+  selectedToken: Token | undefined;
+  onSelectedTokenChange: React.Dispatch<
+    React.SetStateAction<Token | undefined>
+  >;
+  setOpen: React.Dispatch<React.SetStateAction<boolean>>;
+}> = ({
+  isBalancesLoading,
+  balances,
+  selectedToken,
+  onSelectedTokenChange,
+  setOpen,
+}) => {
+  return (
+    <Command
+      defaultValue="-" // to avoid default selected item
+      className="min-h-[340px] max-h-[45svh] bg-transparent border"
+    >
+      <CommandInput placeholder="Search token..." className="text-xs" />
+      <CommandList>
+        {!isBalancesLoading && (
+          <CommandEmpty className="text-xs text-center py-3">
+            No token found.
+          </CommandEmpty>
+        )}
+
+        {isBalancesLoading ? (
+          <div className="p-2">
+            {Array.from({ length: 7 }).map((_, i) => (
+              <div key={i} className="flex items-center gap-2 p-2">
+                <Skeleton className="h-6 min-w-6 rounded-full" />
+                <div className="w-full flex items-center justify-between">
+                  <Skeleton className="h-4 w-14" />
+                  <Skeleton className="h-4 w-24" />
+                </div>
               </div>
-            ) : (
-              <CommandGroup>
-                {balances?.map((token) => {
-                  const key = tokenKey(token);
-                  return (
-                    <TokenListItem
-                      key={key}
-                      itemKey={key}
-                      token={token}
-                      selectedToken={props.token}
-                      setSelectedToken={props.onTokenChange}
-                      setOpen={setOpen}
-                    />
-                  );
-                })}
-              </CommandGroup>
-            )}
-          </CommandList>
-        </Command>
-      </DialogContent>
-    </Dialog>
+            ))}
+          </div>
+        ) : (
+          <CommandGroup>
+            {balances?.map((token) => {
+              const key = tokenKey(token);
+              return (
+                <TokenListItem
+                  key={key}
+                  itemKey={key}
+                  token={token}
+                  selectedToken={selectedToken}
+                  setSelectedToken={onSelectedTokenChange}
+                  setOpen={setOpen}
+                />
+              );
+            })}
+          </CommandGroup>
+        )}
+      </CommandList>
+    </Command>
   );
 };
